+++ modules/codegen/test/src/smithy4s/codegen/internals/TestUtils.scala
@@ -63,6 +63,7 @@ object TestUtils {
   def assertContainsSection(fileContent: String, startsWith: String)(
       expectedSection: String
   )(implicit loc: Location) = {
+    println(fileContent)
     val lines =
       fileContent.linesIterator.filter(_.trim.nonEmpty).zipWithIndex.toList
     val lineMatches = lines.filter { case (l, _) =>
@@ -84,6 +85,34 @@ object TestUtils {
       case _ :: _ :: _ =>
         Assertions.fail("Multiple lines match the code section pattern")
       case Nil => Assertions.fail("No line matches the code section pattern")
+    }
+  }
+
+  /**
+    * Asserts that one of the inputted files contains at least one occurrence of each
+    * of the expectedSections
+    */
+  def assertContainsSection(
+      files: List[String],
+      expectedSections: List[String]
+  )(implicit loc: Location) = {
+    val allFiles = files.mkString("/n")
+
+    expectedSections.foreach { section =>
+      if (!allFiles.contains(section)) {
+        Assertions.fail(
+          "No matching section was found in file contents",
+          new munit.Clues(
+            List(
+              new munit.Clue(
+                source = allFiles,
+                value = section,
+                valueType = "string"
+              )
+            )
+          )
+        )
+      }
     }
   }